Fix performance issue with: GetEntitiesWithTag

+++ src/Sparkle/csharp/Time.cs
@@ -4,7 +4,7 @@ namespace Sparkle.csharp;
 
 public static class Time {
 
-    public static double DeltaTime => Raylib.GetFrameTime();
+    public static float DeltaTime => Raylib.GetFrameTime();
     
     public static double TotalTime => Raylib.GetTime();
